% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_case_detection.R
\name{ActiveCaseDetection.none}
\alias{ActiveCaseDetection.none}
\title{Methods for distributing interventions during active_case_detectional visits}
\usage{
\method{ActiveCaseDetection}{none}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Implements \link{ActiveCaseDetection} for the none model (do nothing)
}
