% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_case_detection.R
\name{ActiveCaseDetection}
\alias{ActiveCaseDetection}
\title{Methods for mass distributing health interventions}
\usage{
ActiveCaseDetection(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a [list]}
}
\value{
[list]
}
\description{
This method dispatches on the type of `pars$active_case_detection`.
}
