% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{add_bednet_events}
\alias{add_bednet_events}
\title{Add Bed Net Rounds}
\usage{
add_bednet_events(xds_obj, jdates, net_type, peak_access)
}
\arguments{
\item{xds_obj}{the \code{ramp.xds} object}

\item{jdates}{the julian date of the bed net mass distribution round}

\item{net_type}{the type of net used}

\item{peak_access}{the fraction of the population with access to a bed net}
}
\value{
a \strong{\code{xds}} object
}
\description{
Add one or more bed net
rounds to the the events list
}
