% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-multiround.R
\name{add_bednet_round}
\alias{add_bednet_round}
\title{Set up dynamic forcing}
\usage{
add_bednet_round(xds_obj, jdate, net_type, peak_access)
}
\arguments{
\item{xds_obj}{the \code{ramp.xds} object}

\item{jdate}{the julian date of the bed net mass distribution round}

\item{net_type}{the type of net used}

\item{peak_access}{the fraction of the population with access to a bed net}
}
\value{
a \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
