% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{add_irs_events}
\alias{add_irs_events}
\title{Add IRS rounds}
\usage{
add_irs_events(xds_obj, jdate, net_type, peak)
}
\arguments{
\item{xds_obj}{the \code{ramp.xds} object}

\item{jdate}{the julian date of the bed net mass distribution round}

\item{net_type}{the type of net used}

\item{peak}{the fraction of houses sprayed}
}
\value{
a \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
