% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_treatment.R
\name{add_mass_treat_events}
\alias{add_mass_treat_events}
\title{Set Up Mass Treatment Events}
\usage{
add_mass_treat_events(xds_obj, jdates, span, frac_tot, test)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{jdates}{the julian dates of mass treatment events}

\item{span}{the treatment period}

\item{frac_tot}{the fraction treated}

\item{test}{FALSE for mass drug administration; TRUE for mass screen and treat}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Set Up Mass Treatment Events
}
