% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-interface.R
\name{AreaSpray.dynamic}
\alias{AreaSpray.dynamic}
\title{Area Spraying}
\usage{
\method{AreaSpray}{dynamic}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} object}
}
\value{
\link{list}
}
\description{
The null model for AreaSpray
}
