% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-none.R
\name{AreaSpray.none}
\alias{AreaSpray.none}
\title{No Area Spraying}
\usage{
\method{AreaSpray}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
[list]
}
\description{
The null model for AreaSpray
}
