% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-interface.R
\name{AreaSpray}
\alias{AreaSpray}
\title{Spray an Area}
\usage{
AreaSpray(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
Set the value of exogenous variables related to
AreaSpray
}
