% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-func.R
\name{AreaSprayCoverage.func}
\alias{AreaSprayCoverage.func}
\title{Set no area_spray_coverage}
\usage{
\method{AreaSprayCoverage}{func}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for area_spray_coverage
}
