% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-none.R
\name{AreaSprayCoverage.none}
\alias{AreaSprayCoverage.none}
\title{Set no AreaSprayCoverage}
\usage{
\method{AreaSprayCoverage}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
[list]
}
\description{
The null model for AreaSprayCoverage
}
