% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-none.R
\name{AreaSprayEffects.none}
\alias{AreaSprayEffects.none}
\title{Set no area spray effects}
\usage{
\method{AreaSprayEffects}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for AreaSprayEffectSizes
}
