% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-none.R
\name{AreaSprayEffectSizes.none}
\alias{AreaSprayEffectSizes.none}
\title{Set no AreaSprayEffectSizes}
\usage{
\method{AreaSprayEffectSizes}{none}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} object}

\item{s}{vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for AreaSprayEffectSizes
}
