% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-interface.R
\name{AreaSprayEffectSizes}
\alias{AreaSprayEffectSizes}
\title{Set the AreaSprayEffectSizes}
\usage{
AreaSprayEffectSizes(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}

\item{s}{vector species index}
}
\value{
an **`xds`** object
}
\description{
Set the value of exogenous variables related to
AreaSprayEffectSizes
}
