% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet.R
\name{Bed_Net}
\alias{Bed_Net}
\title{Set the bednet}
\usage{
Bed_Net(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the value of exogenous variables related to
bednet
}
