% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage.R
\name{Bed_Net_Coverage}
\alias{Bed_Net_Coverage}
\title{Set the cover_bednets}
\usage{
Bed_Net_Coverage(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the value of exogenous variables related to
cover_bednets
}
