% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-effect_sizes.R
\name{Bed_Net_Effect_Sizes}
\alias{Bed_Net_Effect_Sizes}
\title{Bed Net Effect Sizes}
\usage{
Bed_Net_Effect_Sizes(t, y, xds_obj, s = 1)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the \strong{\code{xds}} model object
}
\description{
Compute bed net effect sizes
}
