% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-effects.R
\name{Bed_Net_Effects.none}
\alias{Bed_Net_Effects.none}
\title{No Effects of Bed Nets}
\usage{
\method{Bed_Net_Effects}{none}(t, xds_obj)
}
\arguments{
\item{t}{current time}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the \strong{\code{xds}} model object
}
\description{
Do nothing
}
