% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-effects.R
\name{Bed_Net_Effects}
\alias{Bed_Net_Effects}
\title{Bed Net Effects}
\usage{
Bed_Net_Effects(t, xds_obj)
}
\arguments{
\item{t}{current time}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the \strong{\code{xds}} model object
}
\description{
Set the value of variables in a model that
implements the effects of bed nets
}
