% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-interface.R
\name{BedNet.dynamic}
\alias{BedNet.dynamic}
\title{Set no bednet}
\usage{
\method{BedNet}{dynamic}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
\link{list}
}
\description{
The null model for bednet
}
