% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bednet_profiles}
\alias{bednet_profiles}
\title{Bednet Profiles}
\format{
An object of class \code{data.frame} with 4 rows and 3 columns.
}
\usage{
bednet_profiles
}
\description{
Bednet Profiles
}
\keyword{datasets}
