% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-func.R, R/bednet-functions.R
\name{BedNetCoverage.func}
\alias{BedNetCoverage.func}
\title{Set no bednet_coverage}
\usage{
\method{BedNetCoverage}{func}(t, pars)

\method{BedNetCoverage}{func}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
[list]

[list]
}
\description{
The null model for bednet_coverage

The null model for bednet_coverage
}
