% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-multiround.R
\name{BedNetCoverage.multiround}
\alias{BedNetCoverage.multiround}
\title{Set no bednet_coverage}
\usage{
\method{BedNetCoverage}{multiround}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
The null model for bednet_coverage
}
