% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-interface.R
\name{BedNetCoverage}
\alias{BedNetCoverage}
\title{Set the bednet}
\usage{
BedNetCoverage(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
Set the value of exogenous variables related to
bednet
}
