% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-none.R
\name{BedNetEffects.none}
\alias{BedNetEffects.none}
\title{Set no bednet_effects}
\usage{
\method{BedNetEffects}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
\link{list}
}
\description{
The null model for bednet_effects
}
