% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-interface.R
\name{BedNetEffects}
\alias{BedNetEffects}
\title{Set the bednet_effects}
\usage{
BedNetEffects(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
Set the value of exogenous variables related to
bednet_effects
}
