% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-functions.R
\name{BedNetEffectSizes.func}
\alias{BedNetEffectSizes.func}
\title{Set no bednet}
\usage{
\method{BedNetEffectSizes}{func}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{s}{the vector species index}
}
\value{
\link{list}
}
\description{
The null model for bednet
}
