% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-none.R
\name{BedNetEffectSizes.none}
\alias{BedNetEffectSizes.none}
\title{Set no bednet_effect_sizes}
\usage{
\method{BedNetEffectSizes}{none}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{s}{the vector species index}
}
\value{
\link{list}
}
\description{
The null model for bednet_effect_sizes
}
