% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-interface.R
\name{BedNetEffectSizes}
\alias{BedNetEffectSizes}
\title{Copmute bednet effect sizes}
\usage{
BedNetEffectSizes(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Compute the effect sizes
associated with bed nets
}
