% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/care_seeking.R
\name{CareSeeking}
\alias{CareSeeking}
\title{Set the values of exogenous variables describing care seeking}
\usage{
CareSeeking(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a  \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
This method dispatches on the type of \code{xds_obj$CARE_SEEKING}.
}
