% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-contact-multiround.R
\name{change_bednet_contact_multiround}
\alias{change_bednet_contact_multiround}
\title{Setup Muti-Round bednet Contact}
\usage{
change_bednet_contact_multiround(xds_obj, contact)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{contact}{the new contact parameter}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
With multi-round bednet contact, there
is a different relationship between coverage
and contact in each round
}
