% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-multiround.R
\name{change_bednet_coverage.multiround}
\alias{change_bednet_coverage.multiround}
\title{Set values for bednet coverage}
\usage{
\method{change_bednet_coverage}{multiround}(coverage, xds_obj)
}
\arguments{
\item{coverage}{coverage parameters}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set values for bednet coverage
}
