% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-contact-multiround.R
\name{change_bednet_pw_multiround}
\alias{change_bednet_pw_multiround}
\title{Setup Muti-Round bednet Shape}
\usage{
change_bednet_pw_multiround(xds_obj, pw)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{pw}{the new shape parameter}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
With multi-round bednet pw, there
is a different relationship between coverage
and pw in each round
}
