% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-shock-multiround.R
\name{change_bednet_shock_multiround}
\alias{change_bednet_shock_multiround}
\title{Change bednet Shocks}
\usage{
change_bednet_shock_multiround(xds_obj, shock)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{shock}{the scaling parameter}
}
\value{
an \strong{\code{xds}} model object
}
\description{
For \code{eir} models, set up a multi-round
sharkbite function to model effect sizes
of bednet on the EIR as a perturbation.
}
