% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-contact.R
\name{change_irs_contact}
\alias{change_irs_contact}
\title{Get irs contact}
\usage{
change_irs_contact(contact, xds_obj)
}
\arguments{
\item{contact}{contact parameters}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
maximum irs contact, as a vector
}
\description{
Get the values from a model
that set maximum irs contact
}
