% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-contact-multiround.R
\name{change_irs_contact_multiround}
\alias{change_irs_contact_multiround}
\title{Setup Muti-Round IRS Contact}
\usage{
change_irs_contact_multiround(contact, xds_obj)
}
\arguments{
\item{contact}{contact parameters}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\description{
With multi-round irs contact, there
is a different relationship between coverage
and contact in each round
}
