% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage-multiround.R
\name{change_irs_coverage_multiround}
\alias{change_irs_coverage_multiround}
\title{Setup Muti-Round Bed Net coverage}
\usage{
change_irs_coverage_multiround(coverage, xds_obj)
}
\arguments{
\item{coverage}{the coverage parametter}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\description{
With multi-round bed net coverage, there
is a different relationship between coverage
and coverage in each round
}
