% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic.R
\name{Clinic.none}
\alias{Clinic.none}
\title{Methods for distributing interventions during clinical visits}
\usage{
\method{Clinic}{none}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Implements \link{Clinic} for the none model (do nothing)
}
