% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic.R
\name{Clinic}
\alias{Clinic}
\title{Methods for distributing interventions during clinical visits}
\usage{
Clinic(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
This method dispatches on the type of \code{xds_obj$clinic}.
}
