% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-distribute-none.R
\name{DistributeSugarBaits.none}
\alias{DistributeSugarBaits.none}
\title{Distribute no sugar baits}
\usage{
\method{DistributeSugarBaits}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The model for no sugar bait distribution
}
