% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-distribute-interface.R
\name{DistributeSugarBaits}
\alias{DistributeSugarBaits}
\title{Set the distribute_sugar_baits_sugar_baits}
\usage{
DistributeSugarBaits(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} object}
}
\value{
a \strong{\code{ramp.xds}} object
}
\description{
Set the value of exogenous variables related to
distribute_sugar_baits_sugar_baits
}
