% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{dynamic_health.setup}
\alias{dynamic_health.setup}
\title{Set up dynamic health}
\usage{
\method{dynamic_health}{setup}(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
If dynamic health has not
already been set up, then turn on dynamic
health and set all the
}
