% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{dynamic_vc_effect_sizes}
\alias{dynamic_vc_effect_sizes}
\title{Vector Control Dynamics}
\usage{
dynamic_vc_effect_sizes(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a named \link{list}
}
\description{
Implements various forms
of vector control. Each mode for vector
control is set up and configured separately.
}
\note{
This a junction to implement various modes of
vector control.
Vector control modules require
\href{https://github.com/dd-harp/ramp.control}{\strong{\code{ramp.control}}}.
}
