% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{dynamic_vector_control.dynamic}
\alias{dynamic_vector_control.dynamic}
\title{Vector Control is Turned On}
\usage{
\method{dynamic_vector_control}{dynamic}(pars)
}
\arguments{
\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
If
`class(pars$vector_control) == 'dynamic'`
then dynamic vector control has been turned
on. The unmodified **`xds`** object is returned.
}
