% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-utilities.R
\name{get_bednet_coverage.multiround}
\alias{get_bednet_coverage.multiround}
\title{Set values for bednet coverage}
\usage{
\method{get_bednet_coverage}{multiround}(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
maximum bednet coverage, as a vector
}
\description{
Set values for bednet coverage
}
