% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{get_events.irs}
\alias{get_events.irs}
\title{Get IRS Events}
\usage{
\method{get_events}{irs}(xds_obj, type)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{type}{the event type}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Get IRS Events
}
