% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{get_events}
\alias{get_events}
\title{Get Events}
\usage{
get_events(xds_obj, type)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{type}{the event type}
}
\value{
Events details as a list
}
\description{
Get Events
}
