% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-contact.R
\name{get_irs_contact}
\alias{get_irs_contact}
\title{Get irs contact}
\usage{
get_irs_contact(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Get the values from a model
that set maximum irs contact
}
