% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-utilities.R
\name{get_irs_coverage.multiround}
\alias{get_irs_coverage.multiround}
\title{Set values for irs coverage}
\usage{
\method{get_irs_coverage}{multiround}(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
IRS coverage
}
\description{
Set values for irs coverage
}
