% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-utilities.R
\name{get_irs_coverage}
\alias{get_irs_coverage}
\title{Get irs coverage}
\usage{
get_irs_coverage(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
IRS coverage
}
\description{
Get the values from a model
that set maximum irs coverage
}
