% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-interface.R
\name{IRS}
\alias{IRS}
\title{Implement IRS}
\usage{
IRS(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
Set the value of exogenous variables related to
IRS
}
