% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-contact-func.R
\name{IRS_Contact.func}
\alias{IRS_Contact.func}
\title{Set no irs_contact}
\usage{
\method{IRS_Contact}{func}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for irs_contact
}
