% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-contact-linear.R
\name{IRS_Contact.linear}
\alias{IRS_Contact.linear}
\title{IRS Contact}
\usage{
\method{IRS_Contact}{linear}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up a linear model
for contact relative to coverage
}
