% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-contact.R
\name{IRS_Contact.none}
\alias{IRS_Contact.none}
\title{Set no use_irss}
\usage{
\method{IRS_Contact}{none}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
\link{list}
}
\description{
The null model for use_irss
}
