% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effect_sizes.R
\name{IRS_Effect_Sizes}
\alias{IRS_Effect_Sizes}
\title{IRS effect_sizes}
\usage{
IRS_Effect_Sizes(t, y, xds_obj, s = 1)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the value of variables to model
the effect_sizes of IRS
}
