% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effects.R
\name{IRS_Effects.none}
\alias{IRS_Effects.none}
\title{Set no use_bednets}
\usage{
\method{IRS_Effects}{none}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
\link{list}
}
\description{
The null model for use_bednets
}
