% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irs_profiles}
\alias{irs_profiles}
\title{IRS Profiles}
\format{
An object of class \code{data.frame} with 5 rows and 3 columns.
}
\usage{
irs_profiles
}
\description{
IRS Profiles
}
\keyword{datasets}
