% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage-functions.R
\name{IRSCoverage.func}
\alias{IRSCoverage.func}
\title{Set no irs_coverage}
\usage{
\method{IRSCoverage}{func}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for irs_coverage
}
