% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-none.R
\name{IRSCoverage.none}
\alias{IRSCoverage.none}
\title{Set no irs}
\usage{
\method{IRSCoverage}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for irs
}
