% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-none.R
\name{IRSEffectSizes.none}
\alias{IRSEffectSizes.none}
\title{Set no irs_effectsizes}
\usage{
\method{IRSEffectSizes}{none}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for irs_effectsizes
}
