% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effectsizes-simple.R
\name{IRSEffectSizes.simple}
\alias{IRSEffectSizes.simple}
\title{Modify baseline values due to vector control}
\usage{
\method{IRSEffectSizes}{simple}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{vector species index}
}
\value{
a named \link{list}
}
\description{
Implements \link{IRSEffectSizes} for the Le Menach IRS model of vector control
}
\seealso{
\code{\link[=compute_irs_effect_sizes_simple]{compute_irs_effect_sizes_simple()}}
}
