% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm.R
\name{LSM.dynamic}
\alias{LSM.dynamic}
\title{Set no LSM}
\usage{
\method{LSM}{dynamic}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for LSM
}
