% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-coverage-none.R
\name{LSMCoverage.none}
\alias{LSMCoverage.none}
\title{Set no LSMCoverage}
\usage{
\method{LSMCoverage}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
The null model for LSMCoverage
}
