% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-coverage-interface.R
\name{LSMCoverage}
\alias{LSMCoverage}
\title{Set the LSMCoverage}
\usage{
LSMCoverage(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
Set the value of exogenous variables related to
LSMCoverage
}
