% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-effects-none.R
\name{LSMEffects.none}
\alias{LSMEffects.none}
\title{Set no lsm_effects}
\usage{
\method{LSMEffects}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for lsm_effects
}
