% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-effects-interface.R
\name{LSMEffects}
\alias{LSMEffects}
\title{Set the lsm_effects_lsm}
\usage{
LSMEffects(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the value of exogenous variables related to
lsm_effects_lsm
}
