% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-effectsizes-none.R
\name{LSMEffectSizes.none}
\alias{LSMEffectSizes.none}
\title{Set no LSMEffectSizes}
\usage{
\method{LSMEffectSizes}{none}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for LSMEffectSizes
}
