% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-multiround.R
\name{make_bednet_multiround}
\alias{make_bednet_multiround}
\title{Make the Multiround Bednet Coverage Object}
\usage{
make_bednet_multiround(contact, xds_obj)
}
\arguments{
\item{contact}{a contact parameter}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{xds}} object
}
\description{
Using information about the events (see \link{setup_bednet_events}),
including a parameter describing access, this makes a
function that computes coverage over time.
}
