% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-multiround.R
\name{make_bednet_round.pbo}
\alias{make_bednet_round.pbo}
\title{Set up dynamic forcing}
\usage{
\method{make_bednet_round}{pbo}(type, t_init, coverage, zap = 1)
}
\arguments{
\item{type}{the name of the BedNet type}

\item{t_init}{the time when BedNet started}

\item{coverage}{the coverage achieved}

\item{zap}{contact scaling}
}
\value{
a \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
