% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-multiround.R
\name{make_bednet_round_generic}
\alias{make_bednet_round_generic}
\title{Set up dynamic forcing}
\usage{
make_bednet_round_generic(
  t_init,
  uk = 1/5,
  L = 365,
  dk = 1/60,
  peak = 0.7,
  contact = 1
)
}
\arguments{
\item{t_init}{the time when BedNet started}

\item{uk}{scale up rate}

\item{L}{length of half-life}

\item{dk}{decay shape parameter}

\item{peak}{the peak access achieved}

\item{contact}{contact scaling parameter}
}
\value{
a \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
