% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-multiround.R
\name{make_F_cover_irs}
\alias{make_F_cover_irs}
\title{Set up dynamic forcing}
\usage{
make_F_cover_irs(cover)
}
\arguments{
\item{cover}{a coverage model object}
}
\value{
a function
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
