% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-multiround.R
\name{make_F_multi_bednet}
\alias{make_F_multi_bednet}
\title{Set up dynamic forcing}
\usage{
make_F_multi_bednet(multi_obj)
}
\arguments{
\item{multi_obj}{a list with parameters for all the rounds}
}
\value{
a function
}
\description{
Make a multiround function
}
