% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-multiround.R
\name{make_F_multi_irs}
\alias{make_F_multi_irs}
\title{Set up dynamic forcing}
\usage{
make_F_multi_irs(multi_obj)
}
\arguments{
\item{multi_obj}{a multiround model object}
}
\value{
a function
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
