% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-multiround.R
\name{make_F_multiround_bednet}
\alias{make_F_multiround_bednet}
\title{Set up dynamic forcing}
\usage{
make_F_multiround_bednet(multi_obj, peak, pw = 1)
}
\arguments{
\item{multi_obj}{a multiround model object}

\item{peak}{the scaling parameter}

\item{pw}{a shape parameter}
}
\value{
a function
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
