% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-shock-multiround.R
\name{make_F_multishock_irs}
\alias{make_F_multishock_irs}
\title{Set up dynamic forcing}
\usage{
make_F_multishock_irs(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a function
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
