% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effect_sizes-simple.R
\name{make_irs_effect_sizes_simple}
\alias{make_irs_effect_sizes_simple}
\title{Set up dynamic forcing}
\usage{
make_irs_effect_sizes_simple(options = list(), contact = 1)
}
\arguments{
\item{options}{a list of options to override defaults}

\item{contact}{the probability of contact given coverage}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
A simple model for IRS effect sizes
}
