% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-multiround.R
\name{make_irs_multiround}
\alias{make_irs_multiround}
\title{Set up dynamic forcing}
\usage{
make_irs_multiround(xds_obj, use_contact = TRUE)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{use_contact}{TRUE uses contact parameters from the events object}
}
\value{
an IRS coverage model object
}
\description{
Set up a function that computes
irs coverage over time.
}
