% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-pesticides.R
\name{make_irs_round}
\alias{make_irs_round}
\title{Set up dynamic forcing}
\usage{
make_irs_round(type, t_init, coverage, contact)
}
\arguments{
\item{type}{the name of the IRS type}

\item{t_init}{the time when IRS started}

\item{coverage}{effective coverage}

\item{contact}{effective contact}
}
\value{
a \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
