% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_treatment.R
\name{make_mass_treat_multiround}
\alias{make_mass_treat_multiround}
\title{Make Multiple Rounds of msat}
\usage{
make_mass_treat_multiround(xds_obj, screen)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{screen}{a switch}
}
\value{
a \strong{\code{xds}} object
}
\description{
Using information about the events (see \link{setup_mass_treat_events}),
including a parameter describing access, this makes a
function that computes treatage over time.
}
