% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-functions.R
\name{OwnBedNets.func}
\alias{OwnBedNets.func}
\title{Set no own_bednets}
\usage{
\method{OwnBedNets}{func}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
[list]
}
\description{
The null model for own_bednets
}
