% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school.R
\name{School}
\alias{School}
\title{Methods for distributing interventions during schoolal visits}
\usage{
School(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
This method dispatches on the type of \code{xds_obj$school}.
}
