% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-utilities.R
\name{set_bednet_coverage}
\alias{set_bednet_coverage}
\title{Get bednet coverage}
\usage{
set_bednet_coverage(coverage, xds_obj)
}
\arguments{
\item{coverage}{coverage parameters}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
maximum bednet coverage, as a vector
}
\description{
Get the values from a model
that set maximum bednet coverage
}
