% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-utilities.R
\name{set_irs_coverage}
\alias{set_irs_coverage}
\title{Get irs coverage}
\usage{
set_irs_coverage(coverage, xds_obj)
}
\arguments{
\item{coverage}{coverage parameters}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Get the values from a model
that set maximum irs coverage
}
