% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-interface.R
\name{setup_area_spray}
\alias{setup_area_spray}
\title{Set up area spraying}
\usage{
setup_area_spray(
  xds_obj,
  spray_area_name = "none",
  spray_area_opts = list(),
  effects_name = "none",
  effects_opts = list(),
  coverage_name = "none",
  coverage_opts = list(),
  effectsizes_name = "none",
  effectsizes_opts = list()
)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} object}

\item{spray_area_name}{the name of a model for mass bed net distribution}

\item{spray_area_opts}{options for the bed net distribution model}

\item{effects_name}{the name of a model for bed net effects}

\item{effects_opts}{options for the bed net effects model}

\item{coverage_name}{the name of a model for bed net coverage}

\item{coverage_opts}{options for the bed net coverage model}

\item{effectsizes_name}{the name of a model for bed net effect sizes}

\item{effectsizes_opts}{options for the bed net effect sizes model}
}
\value{
a \strong{\code{ramp.xds}} object
}
\description{
If the dynamic switch has not
already been set for Area Spraying, turn it on
and set up the modules.
}
