% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-none.R
\name{setup_area_spray_coverage.none}
\alias{setup_area_spray_coverage.none}
\title{Set up no area spray}
\usage{
\method{setup_area_spray_coverage}{none}(name = "none", xds_obj, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}} object}

\item{opts}{a list of options to override defaults}
}
\value{
a \strong{\code{xds}} object
}
\description{
Set up no area spray
}
