% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-none.R
\name{setup_area_spray_effect_sizes.none}
\alias{setup_area_spray_effect_sizes.none}
\title{Set up no area spray}
\usage{
setup_area_spray_effect_sizes.none(name = "none", pars, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an **`xds`** object}

\item{opts}{a list of options to override defaults}
}
\value{
an **`xds`** object
}
\description{
Set up no area spray
}
