% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-interface.R
\name{setup_area_spray_effects}
\alias{setup_area_spray_effects}
\title{Set up dynamic area spray}
\usage{
setup_area_spray_effects(name, xds_obj, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}} object}

\item{opts}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} object
}
\description{
If dynamic area spray has not
already been set up, then turn on dynamic
area spray and set all the
}
