% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray.R
\name{setup_area_spray_object}
\alias{setup_area_spray_object}
\title{Set Up the Null Area Spray Model Object}
\usage{
setup_area_spray_object(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Sets up the null area spray model object
}
