% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_baits.R
\name{setup_atsb_object}
\alias{setup_atsb_object}
\title{Set Up the Null Sugar Baits Model Object}
\usage{
setup_atsb_object(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Sets up the null sugar baits model object
}
