% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-contact.R
\name{setup_bednet_contact}
\alias{setup_bednet_contact}
\title{Set Up the Bed Net Contact Object}
\usage{
setup_bednet_contact(name, xds_obj, options = list())
}
\arguments{
\item{name}{the model name}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{options}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up an object to model
bed net effective contact.
}
