% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-func.R, R/bednet-functions.R
\name{setup_bednet_coverage_func}
\alias{setup_bednet_coverage_func}
\title{Set up dynamic forcing}
\usage{
setup_bednet_coverage_func(
  pars,
  opts = list(),
  mean = 1,
  F_season = F_flat,
  season_par = list(),
  F_trend = F_flat,
  trend_par = list()
)

setup_bednet_coverage_func(
  pars,
  opts = list(),
  mean = 1,
  F_season = F_flat,
  season_par = list(),
  F_trend = F_flat,
  trend_par = list()
)
}
\arguments{
\item{pars}{an **`xds`** object}

\item{opts}{a list of options to override defaults}

\item{mean}{the mean bednet_coverage}

\item{F_season}{the seasonal signal in bednet coverage}

\item{season_par}{parameters to configure F_season}

\item{F_trend}{a temporal trend in bednet coverage}

\item{trend_par}{parameters to configure F_trend}
}
\value{
an **`xds`** object

an **`xds`** object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the

If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
