% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-interface.R
\name{setup_bednet_effects}
\alias{setup_bednet_effects}
\title{Set up dynamic forcing}
\usage{
setup_bednet_effects(name, xds_obj, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{opts}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
