% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-functions.R
\name{setup_bednet_effectsizes.func}
\alias{setup_bednet_effectsizes.func}
\title{Set up dynamic forcing}
\usage{
<<<<<<< HEAD
\method{setup_bednet_effectsizes}{func}(name, pars, opts = list())
=======
\method{setup_bednet_effectsizes}{func}(name, pars, s = 1, opts = list())
>>>>>>> db0563d1faeb889963b6d07f2c68be34e436ff93
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an **`xds`** object}

\item{s}{the vector species index}

\item{opts}{a list of options to override defaults}
}
\value{
an **`xds`** object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
