% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-interface.R
\name{setup_bednet_effectsizes}
\alias{setup_bednet_effectsizes}
\title{Set up dynamic forcing}
\usage{
setup_bednet_effectsizes(name, pars, s = 1, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an **`xds`** object}

\item{s}{the vector species index}

\item{opts}{a list of options to override defaults}
}
\value{
an **`xds`** object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
