% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-multiround.R
\name{setup_bednet_multiround}
\alias{setup_bednet_multiround}
\title{Set up dynamic forcing}
\usage{
setup_bednet_multiround(
  opts = list(),
  t_init = 1,
  coverage = 0.8,
  type = "pbo",
  zap = 1
)
}
\arguments{
\item{opts}{a list of options to override defaults}

\item{t_init}{the time when BedNet started}

\item{coverage}{the coverage achieved}

\item{type}{the BedNet type}

\item{zap}{the contact parameter}
}
\value{
an **`xds`** object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
