% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet.R
\name{setup_bednet_object}
\alias{setup_bednet_object}
\title{Set Up the Null Bednet Model Object}
\usage{
setup_bednet_object(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Sets up the null bednet model object
}
