% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-own.R
\name{setup_bednet_own}
\alias{setup_bednet_own}
\title{Set Up the Bed Net Ownership Object}
\usage{
setup_bednet_own(name, xds_obj, options = list())
}
\arguments{
\item{name}{the model name}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{options}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up an object to model
bed net ownership
}
