% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-multiround.R
\name{setup_bednet_rounds}
\alias{setup_bednet_rounds}
\title{Set up dynamic forcing}
\usage{
setup_bednet_rounds(xds_obj, mx, as_shock = FALSE)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{mx}{the maximum}

\item{as_shock}{set the function class to \code{sharkfin} (if \code{FALSE}) or \code{sharkbite} (if \code{TRUE})}
}
\value{
set up the rounds
}
\description{
Set up the bednet rounds
}
