% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-shock-multiround.R
\name{setup_bednet_shock_multiround}
\alias{setup_bednet_shock_multiround}
\title{Setup Bed Net Shocks}
\usage{
setup_bednet_shock_multiround(xds_obj, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}}  model object}

\item{options}{a list of options}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Set up a function that computes
bednet multi_objage over time.
}
