% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-functions.R
\name{setup_distribute_bednets.func}
\alias{setup_distribute_bednets.func}
\title{Set up dynamic bednets}
\usage{
\method{setup_distribute_bednets}{func}(name = "func", xds_obj, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{opts}{a list of options to override defaults}
}
\description{
If dynamic bednets has not
already been set up, then turn on dynamic
bednets and set all the
}
