% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-interface.R
\name{setup_distribute_bednets}
\alias{setup_distribute_bednets}
\title{Set up dynamic bednets}
\usage{
setup_distribute_bednets(name, xds_obj, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{opts}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
If dynamic bednets has not
already been set up, then turn on dynamic
bednets and set all the
}
