% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-multiround.R
\name{setup_F_cover_bednet}
\alias{setup_F_cover_bednet}
\title{Set up dynamic forcing}
\usage{
setup_F_cover_bednet(cover)
}
\arguments{
\item{cover}{a list with parameters for all the rounds}
}
\value{
a \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
