% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage-multiround.R
\name{setup_F_cover_irs}
\alias{setup_F_cover_irs}
\title{Set up dynamic forcing}
\usage{
setup_F_cover_irs(cover)
}
\arguments{
\item{cover}{a mo}

\item{t_init}{the time when irs started}

\item{coverage}{the coverage achieved}

\item{type}{the irs type}

\item{zap}{the contact parameter}
}
\value{
an **`xds`** object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
