% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage-multiround.R
\name{setup_F_coverage_irs_multiround}
\alias{setup_F_coverage_irs_multiround}
\title{Make \code{F_coverage} for IRS}
\usage{
setup_F_coverage_irs_multiround(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
set up the rounds
}
\description{
Set up the IRS rounds
}
