% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-contact-func.R
\name{setup_irs_contact.func}
\alias{setup_irs_contact.func}
\title{Set Up a IRS Contact Function}
\usage{
\method{setup_irs_contact}{func}(name = "func", xds_obj, options = list())
}
\arguments{
\item{name}{the model name}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{options}{a list of options to override defaults}
}
\description{
This sets up a function to set the value of bed net
contact
}
