% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-contact-linear.R
\name{setup_irs_contact.linear}
\alias{setup_irs_contact.linear}
\title{Set Up the Null IRS Contact Object}
\usage{
\method{setup_irs_contact}{linear}(name, xds_obj, options = list(cp = 1))
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{options}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up the null irs coverage object
}
