% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-contact-multiround.R
\name{setup_irs_contact.multiround}
\alias{setup_irs_contact.multiround}
\title{Set up dynamic forcing}
\usage{
\method{setup_irs_contact}{multiround}(name, xds_obj, options = list())
}
\arguments{
\item{name}{the model name}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{options}{a list of options to override defaults}
}
\value{
a IRS contact model object
}
\description{
A set up utility function for
\link{IRS_Contact}.
}
