% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-contact.R
\name{setup_irs_contact}
\alias{setup_irs_contact}
\title{Set Up the IRS Contact Object}
\usage{
setup_irs_contact(name, xds_obj, options = list())
}
\arguments{
\item{name}{the model name}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{options}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up an object to model
the effects of IRS, usually through the
computation of some intermediate variable
}
