% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-none.R
\name{setup_irs_effect_sizes.none}
\alias{setup_irs_effect_sizes.none}
\title{Set up "no irs_effectsizes"}
\usage{
\method{setup_irs_effect_sizes}{none}(name, pars, opts)
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an **`xds`** object}

\item{opts}{a list of options to override defaults}
}
\value{
an irs effect size model object
}
\description{
Set up "no irs_effectsizes"
}
