% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effect_sizes-simple.R
\name{setup_irs_effect_sizes.simple}
\alias{setup_irs_effect_sizes.simple}
\title{Set up dynamic forcing}
\usage{
\method{setup_irs_effect_sizes}{simple}(name, xds_obj, s = 1, options = list())
}
\arguments{
\item{name}{the model name}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{s}{the vector species index}

\item{options}{a list of options to override defaults}
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
