% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effectsizes-simple.R
\name{setup_irs_effect_sizes_simple}
\alias{setup_irs_effect_sizes_simple}
\title{Set up dynamic forcing}
\usage{
setup_irs_effect_sizes_simple(xds_obj, opts = list(), contact = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{opts}{a list of options to override defaults}

\item{contact}{the probability of contact given coverage}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
