% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-simple.R
\name{setup_irs_effectsizes.simple}
\alias{setup_irs_effectsizes.simple}
\title{Set up dynamic forcing}
\usage{
\method{setup_irs_effectsizes}{simple}(name, pars, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an **`xds`** object}

\item{opts}{a list of options to override defaults}
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
