% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effectsizes-simple.R
\name{setup_irs_effectsizes_simple}
\alias{setup_irs_effectsizes_simple}
\title{Set up dynamic forcing}
\usage{
setup_irs_effectsizes_simple(pars, s = 1, opts = list(), contact = 1)
}
\arguments{
\item{pars}{an **`xds`** object}

\item{s}{the vector species index}

\item{opts}{a list of options to override defaults}

\item{contact}{the probability of contact given coverage}
}
\value{
an **`xds`** object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
