% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-pesticides.R
\name{setup_irs_round.none}
\alias{setup_irs_round.none}
\title{Set up dynamic forcing}
\usage{
\method{setup_irs_round}{none}(type, t_init, coverage, zap = 1)
}
\arguments{
\item{type}{the name of the IRS type}

\item{t_init}{the time when IRS started}

\item{coverage}{the coverage achieved}

\item{zap}{contact scaling}
}
\value{
a \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
