% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage-multiround.R
\name{setup_irs_round_generic}
\alias{setup_irs_round_generic}
\title{Set up dynamic forcing}
\usage{
setup_irs_round_generic(
  t_init,
  uk = 1/5,
  L = 365,
  dk = 1/60,
  coverage = 0.7,
  zap = 1
)
}
\arguments{
\item{t_init}{the time when IRS started}

\item{uk}{scale up rate}

\item{L}{length of half-life}

\item{dk}{decay shape parameter}

\item{coverage}{the coverage achieved}

\item{zap}{contact scaling parameter}
}
\value{
an **`xds`** object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
