% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-shock-multiround.R
\name{setup_irs_shock_multiround}
\alias{setup_irs_shock_multiround}
\title{Setup IRS Shocks}
\usage{
setup_irs_shock_multiround(xds_obj, options = list())
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{options}{a list of options}
}
\value{
an IRS multi_objage / contact model object
}
\description{
Set up a function that computes
irs multi_objage over time.
}
