% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-coverage-none.R
\name{setup_lsm_coverage.none}
\alias{setup_lsm_coverage.none}
\title{Set up "no lsm_coverage"}
\usage{
\method{setup_lsm_coverage}{none}(name = "none", xds_obj, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{opts}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up "no lsm_coverage"
}
