% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-effects-none.R
\name{setup_lsm_effects.none}
\alias{setup_lsm_effects.none}
\title{Set up "no lsm_effects"}
\usage{
\method{setup_lsm_effects}{none}(name, xds_obj, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{opts}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up "no lsm_effects"
}
