% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-effects-interface.R
\name{setup_lsm_effects}
\alias{setup_lsm_effects}
\title{Set up dynamic lsm}
\usage{
setup_lsm_effects(name, xds_obj, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{opts}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
If dynamic lsm has not
already been set up, then turn on dynamic
lsm and set all the
}
