% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm.R
\name{setup_lsm_object}
\alias{setup_lsm_object}
\title{Set Up the Null LSM Model Object}
\usage{
setup_lsm_object(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Sets up the null LSM model object
}
