% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-none.R
\name{setup_no_bednets}
\alias{setup_no_bednets}
\title{Set up "no bednet"}
\usage{
setup_no_bednets(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up "no bednet"
}
