% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-none.R
\name{setup_no_irs}
\alias{setup_no_irs}
\title{Set up "no irs"}
\usage{
setup_no_irs(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up "no irs"
}
