% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm.R
\name{setup_no_lsm}
\alias{setup_no_lsm}
\title{Set up "no LSM"}
\usage{
setup_no_lsm(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up "no LSM"
}
