% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school.R
\name{setup_no_school}
\alias{setup_no_school}
\title{Set up the none model for schoolal distribution (do nothing)}
\usage{
setup_no_school(xds_obj)
}
\arguments{
\item{xds_obj}{a a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up the none model for schoolal distribution (do nothing)
}
