% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait.R
\name{setup_no_sugar_baits}
\alias{setup_no_sugar_baits}
\title{Set up no sugar baits}
\usage{
setup_no_sugar_baits(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up no sugar baits
}
