% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-none.R
\name{setup_own_bednets.none}
\alias{setup_own_bednets.none}
\title{Set up "no own_bednets"}
\usage{
\method{setup_own_bednets}{none}(name, xds_obj, opts)
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{opts}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up "no own_bednets"
}
