% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-none.R
\name{setup_spray_houses.none}
\alias{setup_spray_houses.none}
\title{Set up "no spray_houses"}
\usage{
\method{setup_spray_houses}{none}(name, xds_obj, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{opts}{a list of options to override defaults}
}
\value{
a irs spray model object
}
\description{
Set up "no spray_houses"
}
