% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-spray_houses.R
\name{setup_spray_houses}
\alias{setup_spray_houses}
\title{Set Up the House Spraying Object}
\usage{
setup_spray_houses(name, xds_obj, options = list())
}
\arguments{
\item{name}{the model name}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{options}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up an object to model
house spraying
}
