% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-effects-none.R
\name{setup_sugar_bait_effects.none}
\alias{setup_sugar_bait_effects.none}
\title{Set up no sugar bait effects}
\usage{
\method{setup_sugar_bait_effects}{none}(name, xds_obj, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{opts}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up no sugar bait effects
}
