% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar-bait-effectsizes-none.R
\name{setup_sugar_bait_effectsizes.none}
\alias{setup_sugar_bait_effectsizes.none}
\title{Set up "no sugar_bait_effectsizes"}
\usage{
\method{setup_sugar_bait_effectsizes}{none}(name, xds_obj, opts)
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{opts}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up "no sugar_bait_effectsizes"
}
