% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{setup_vector_control}
\alias{setup_vector_control}
\title{Set up Vector Control}
\usage{
setup_vector_control(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Any function that sets up
vector control calls this function.

The cases are:
\itemize{
\item \code{none} the object needs to be initialized
\item \code{done} the object has been initialized
}

If any function sets \code{vector_control_obj} to \code{dynamic} then
every other
trivial module for every mode of
vector control is set up. Otherwise,
nothing happens.
}
