% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{setup_vector_control.setup}
\alias{setup_vector_control.setup}
\title{Vector Control is in Setup Mode}
\usage{
\method{setup_vector_control}{setup}(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
If \code{class(xds_obj$vector_control) == 'setup'}
then some other mode of vector control has already
changed the setup. Nothing needs to be changed.
The unmodified \strong{\code{ramp.xds}} model object is returned.
}
