% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{setup_vector_control.static}
\alias{setup_vector_control.static}
\title{Vector Control is Static}
\usage{
\method{setup_vector_control}{static}(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
If \code{class(xds_obj$vector_control) == 'static'}
then vector control has already been set up.

This changes the class to to 'setup' to force
the model to run \code{dynamic_vector_control} once.
}
