% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-contact.R
\name{show_bednet_contact}
\alias{show_bednet_contact}
\title{Plot Bednet Contact}
\usage{
show_bednet_contact(tt, xds_obj, clr = "black", add = FALSE)
}
\arguments{
\item{tt}{a set of time points}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{clr}{plotting color}

\item{add}{add to an existing plot}
}
\value{
a \strong{\code{ramp.xds}} xds_obj object
}
\description{
Plot Bednet Contact
}
