% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-profiles.R
\name{show_bednet_response_timeline}
\alias{show_bednet_response_timeline}
\title{Make Parameters for an bednet Round}
\usage{
show_bednet_response_timeline(coverage_profile)
}
\arguments{
\item{coverage_profile}{the name of the bednet type}
}
\value{
a \strong{\code{xds}} object
}
\description{
Return the parameters
to call a sharkfin function to model
a single round of bednet.
}
