% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{show_coverage}
\alias{show_coverage}
\title{Plot coverage}
\usage{
show_coverage(tt, model, mode, add = FALSE)
}
\arguments{
\item{tt}{a set of time points}

\item{model}{a **`ramp.xds`** model object}

\item{mode}{a control mode}

\item{add}{add to an existing plot}
}
\value{
a **`ramp.xds`** model object
}
\description{
Plot coverage
}
