% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-profiles.R
\name{show_coverage_profile}
\alias{show_coverage_profile}
\title{Show an bednet Profile}
\usage{
show_coverage_profile(coverage_profile)
}
\arguments{
\item{coverage_profile}{the name of the bednet type}
}
\value{
a \strong{\code{xds}} object
}
\description{
Return the parameters
to call a sharkfin function to model
a single round of bednet.
}
