% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_treatment.R
\name{show_mda}
\alias{show_mda}
\title{Show MDA}
\usage{
show_mda(tt, xds_obj, clr = "black", add = FALSE)
}
\arguments{
\item{tt}{a set of time points}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{add}{if FALSE, use plot to draw new axes}

\item{clrs}{the line color(s)}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Plot the per-capita mass treatment rate
}
