% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-none.R
\name{SprayArea.none}
\alias{SprayArea.none}
\title{Set no spray_area}
\usage{
\method{SprayArea}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for spray_area
}
