% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-interface.R
\name{SprayHouses}
\alias{SprayHouses}
\title{Set the spray_houses_irs}
\usage{
SprayHouses(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the value of exogenous variables related to
spray_houses_irs
}
