% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-coverage-func.R
\name{SugarBaitCoverage.func}
\alias{SugarBaitCoverage.func}
\title{Set up sugar bait coverage function}
\usage{
\method{SugarBaitCoverage}{func}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{xds}} object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up sugar bait coverage function
}
