% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-coverage-none.R
\name{SugarBaitCoverage.none}
\alias{SugarBaitCoverage.none}
\title{No sugar baits}
\usage{
\method{SugarBaitCoverage}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{xds}} object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for sugar baits
}
