% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-coverage-interface.R
\name{SugarBaitCoverage}
\alias{SugarBaitCoverage}
\title{Set the SugarBaitCoverage}
\usage{
SugarBaitCoverage(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
Set the value of exogenous variables related to
Sugar Baits
}
