% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-effects-none.R
\name{SugarBaitEffects.none}
\alias{SugarBaitEffects.none}
\title{Set up no sugar bait effects}
\usage{
\method{SugarBaitEffects}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up no sugar bait effects
}
