% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-effects-interface.R
\name{SugarBaitEffects}
\alias{SugarBaitEffects}
\title{Compute sugar bait effects}
\usage{
SugarBaitEffects(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Compute the effects of a sugar bait
}
