% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait.R
\name{SugarBaits.dynamic}
\alias{SugarBaits.dynamic}
\title{Set no sugar baits}
\usage{
\method{SugarBaits}{dynamic}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
\link{list}
}
\description{
The null model for sugar baits
}
