% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait.R
\name{SugarBaits}
\alias{SugarBaits}
\title{Set the SugarBaits}
\usage{
SugarBaits(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the value of exogenous variables related to
Sugar Baits
}
