% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-treat_habitats-none.R
\name{TreatHabitats.none}
\alias{TreatHabitats.none}
\title{Set no treat_habitats}
\usage{
\method{TreatHabitats}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for treat_habitats
}
