% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-functions.R
\name{UseBedNets.func}
\alias{UseBedNets.func}
\title{Set no use_bednets}
\usage{
\method{UseBedNets}{func}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
\link{list}
}
\description{
The null model for use_bednets
}
