% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-interface.R
\name{UseBedNets}
\alias{UseBedNets}
\title{Set the use_bednets}
\usage{
UseBedNets(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
Set the value of exogenous variables related to
use_bednets
}
