% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{VectorControl.dynamic}
\alias{VectorControl.dynamic}
\title{Dynamic Vector Control}
\usage{
\method{VectorControl}{dynamic}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a named \link{list}
}
\description{
Calls \code{dynamic_vector_control}
}
\note{
This the function that implements various modes of
vector control.
}
