% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{VectorControl.setup}
\alias{VectorControl.setup}
\title{Vector Control for Static Vector Control}
\usage{
\method{VectorControl}{setup}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The \code{setup} case runs the \code{dynamic_vector_control}
and \code{dynamic_vc_effect_sizes} once, to ensure
that everything is configured
properly for static models. It then changes the
case back to \code{static}

This calls both functions, so there
is no \code{setup} method for \code{VectorControlEffectSizes}
is needed
}
