% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{VectorControl2.static}
\alias{VectorControl2.static}
\title{Distribute vector control, the null model}
\usage{
\method{VectorControl2}{static}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a named \link{list}
}
\description{
Implements \link[ramp.xds:VectorControl1]{ramp.xds::VectorControl1} for the control model of vector control (do nothing)
}
