% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{VectorControlEffectSizes.dynamic}
\alias{VectorControlEffectSizes.dynamic}
\title{Distribute vector control, the null model}
\usage{
\method{VectorControlEffectSizes}{dynamic}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}
}
\value{
a named \link{list}
}
\description{
Implements \link{VectorControl} for the control model of vector control (do nothing)
}
