% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{VectorControlEffectSizes.static}
\alias{VectorControlEffectSizes.static}
\title{Distribute vector control, the null model}
\usage{
\method{VectorControlEffectSizes}{static}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a named \link{list}
}
\description{
Implements \link[ramp.xds:VectorControl]{ramp.xds::VectorControl} for the control model of vector control (do nothing)
}
