% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_case_detection.R
\name{ActiveCaseDetection}
\alias{ActiveCaseDetection}
\title{Methods for mass distributing health interventions}
\usage{
ActiveCaseDetection(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
This method dispatches on the type of \code{xds_obj$active_case_detection}.
}
