% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{add_bednet_events}
\alias{add_bednet_events}
\title{Add Bed Net Rounds}
\usage{
add_bednet_events(
  xds_obj,
  start_day,
  net_type,
  coverage_profile,
  peak_access,
  event_length = 20
)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{start_day}{the julian dates of bed net mass distribution events}

\item{net_type}{the type of net used}

\item{coverage_profile}{the type of net used}

\item{peak_access}{the fraction of the population with access to a bed net}

\item{event_length}{the length of the bed net distribution event}
}
\value{
a \strong{\code{xds}} object
}
\description{
Add one or more bed net
rounds to the the events list
}
