% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-multiround.R
\name{add_bednet_round}
\alias{add_bednet_round}
\title{Set up dynamic forcing}
\usage{
add_bednet_round(xds_obj, type, t_init, coverage, zap = 1)
}
\arguments{
\item{xds_obj}{the \code{ramp.xds} object}

\item{type}{the name of the BedNet type}

\item{t_init}{the time when BedNet started}

\item{coverage}{the coverage achieved}

\item{zap}{the coverage achieved}
}
\value{
a \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
