% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{add_irs_events}
\alias{add_irs_events}
\title{Add IRS rounds}
\usage{
add_irs_events(xds_obj, start_day, pesticides, frac_sprayed, event_length = 20)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{start_day}{the Julian start dates of IRS events}

\item{pesticides}{the pesticide used}

\item{frac_sprayed}{the fraction of houses sprayed}

\item{event_length}{the length of the distribution event}
}
\value{
a \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
