% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{add_irs_events}
\alias{add_irs_events}
\title{Add IRS rounds}
\usage{
add_irs_events(xds_obj, jdates, pesticides, frac_sprayed, contact)
}
\arguments{
\item{xds_obj}{the \code{ramp.xds} object}

\item{jdates}{the julian dates of IRS events}

\item{pesticides}{the pesticide used}

\item{frac_sprayed}{the fraction of houses sprayed}

\item{contact}{the contact parameter}
}
\value{
a \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
