% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-interface.R
\name{AreaSpray.dynamic}
\alias{AreaSpray.dynamic}
\title{Area Spraying}
\usage{
\method{AreaSpray}{dynamic}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
[list]
}
\description{
The null model for AreaSpray
}
