% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-none.R
\name{AreaSpray.none}
\alias{AreaSpray.none}
\title{No Area Spraying}
\usage{
\method{AreaSpray}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for AreaSpray
}
