% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-interface.R
\name{AreaSpray}
\alias{AreaSpray}
\title{Spray an Area}
\usage{
AreaSpray(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} object}
}
\value{
a \strong{\code{ramp.xds}} object
}
\description{
Set the value of exogenous variables related to
AreaSpray
}
