% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-none.R
\name{AreaSprayCoverage.none}
\alias{AreaSprayCoverage.none}
\title{Set no AreaSprayCoverage}
\usage{
\method{AreaSprayCoverage}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for AreaSprayCoverage
}
