% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-interface.R
\name{AreaSprayEffectSizes}
\alias{AreaSprayEffectSizes}
\title{Set the AreaSprayEffectSizes}
\usage{
AreaSprayEffectSizes(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} object}

\item{s}{vector species index}
}
\value{
a \strong{\code{ramp.xds}} object
}
\description{
Set the value of exogenous variables related to
AreaSprayEffectSizes
}
