% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-contact-multiround.R
\name{Bed_Net_Contact.multiround}
\alias{Bed_Net_Contact.multiround}
\title{Set no bednet_coverage}
\usage{
\method{Bed_Net_Contact}{multiround}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{xds}} object
}
\description{
The null model for bednet_coverage
}
