% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-functions.R
\name{BedNetCoverage.func}
\alias{BedNetCoverage.func}
\title{Set no bednet_coverage}
\usage{
\method{BedNetCoverage}{func}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
\link{list}
}
\description{
The null model for bednet_coverage
}
