% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-multiround.R
\name{BedNetCoverage.multiround}
\alias{BedNetCoverage.multiround}
\title{Set no bednet_coverage}
\usage{
\method{BedNetCoverage}{multiround}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{xds}} object
}
\description{
The null model for bednet_coverage
}
