% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-none.R
\name{BedNetEffects.none}
\alias{BedNetEffects.none}
\title{Set no bednet_effects}
\usage{
\method{BedNetEffects}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
[list]
}
\description{
The null model for bednet_effects
}
