% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-functions.R
\name{BedNetEffectSizes.func}
\alias{BedNetEffectSizes.func}
\title{Set no bednet}
\usage{
\method{BedNetEffectSizes}{func}(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}

\item{s}{the vector species index}
}
\value{
[list]
}
\description{
The null model for bednet
}
