% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-interface.R
\name{BedNetEffectSizes}
\alias{BedNetEffectSizes}
\title{Copmute bednet effect sizes}
\usage{
BedNetEffectSizes(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}

\item{s}{the vector species index}
}
\value{
an **`xds`** object
}
\description{
Compute the effect sizes
associated with bed nets
}
