% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/care_seeking.R
\name{CareSeeking}
\alias{CareSeeking}
\title{Set the values of exogenous variables describing care seeking}
\usage{
CareSeeking(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a [list]}
}
\value{
[list]
}
\description{
This method dispatches on the type of `pars$CARE_SEEKING`.
}
