% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-contact-multiround.R
\name{change_bednet_contact_multiround}
\alias{change_bednet_contact_multiround}
\title{Setup Muti-Round Bed Net Contact}
\usage{
change_bednet_contact_multiround(contact, xds_obj)
}
\arguments{
\item{contact}{the contact parametter}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\description{
With multi-round bed net contact, there
is a different relationship between coverage
and contact in each round
}
