% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-multiround.R
\name{change_bednet_contact_multiround}
\alias{change_bednet_contact_multiround}
\title{Set values for bednet coverage}
\usage{
change_bednet_contact_multiround(contact, xds_obj)
}
\arguments{
\item{contact}{contact parameter(s)}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set values for bednet coverage
}
