% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage-multiround.R
\name{change_irs_coverage_multiround}
\alias{change_irs_coverage_multiround}
\title{Change IRS Multiround Coverage}
\usage{
change_irs_coverage_multiround(xds_obj, frac_sprayed)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{frac_sprayed}{the fraction of houses sprayed}
}
\description{
With multi-round bed net coverage, there
is a different relationship between coverage
and coverage in each round
}
