% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic.R
\name{Clinic}
\alias{Clinic}
\title{Methods for distributing interventions during clinical visits}
\usage{
Clinic(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a [list]}
}
\value{
[list]
}
\description{
This method dispatches on the type of `pars$clinic`.
}
