% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-simple.R
\name{compute_irs_effect_sizes_simple}
\alias{compute_irs_effect_sizes_simple}
\title{Modify baseline values due to vector control}
\usage{
compute_irs_effect_sizes_simple(ix, phi, ff, qq, gg, contact = 0.8)
}
\arguments{
\item{ix}{an index over nPatches}

\item{phi}{IRS coverage}

\item{ff}{baseline blood feeding rate}

\item{qq}{baseline human fraction}

\item{gg}{baseline mosquito mortality rate}

\item{contact}{the probability of contact given coverage}
}
\value{
an **`xds`** model object
}
\description{
Implements a model published in 2007
}
