% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-none.R
\name{DistributeBedNets.none}
\alias{DistributeBedNets.none}
\title{Set no distribute_bednets}
\usage{
\method{DistributeBedNets}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
\link{list}
}
\description{
The null model for distribute_bednets
}
