% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{dynamic_vector_control.dynamic}
\alias{dynamic_vector_control.dynamic}
\title{Vector Control is Turned On}
\usage{
\method{dynamic_vector_control}{dynamic}(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
If
\code{class(xds_obj$vector_control) == 'dynamic'}
then dynamic vector control has been turned
on. The unmodified \strong{\code{ramp.xds}} model object is returned.
}
