% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-multiround.R
\name{get_bednet_contact}
\alias{get_bednet_contact}
\title{Get Bednet Contact Parameter}
\usage{
get_bednet_contact(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
the contact parameter
}
\description{
Get Bednet Contact Parameter
}
