% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage-multiround.R
\name{IRSCoverage.multiround}
\alias{IRSCoverage.multiround}
\title{IRS Coverage with Multiple Rounds}
\usage{
\method{IRSCoverage}{multiround}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
A model for IRS coverage over time
when there have been several, different IRS models
}
