% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effectsizes-simple.R
\name{IRSEffectSizes.simple}
\alias{IRSEffectSizes.simple}
\title{Modify baseline values due to vector control}
\usage{
\method{IRSEffectSizes}{simple}(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{vector species index}
}
\value{
a named [list]
}
\description{
Implements [IRSEffectSizes] for the Le Menach IRS model of vector control
}
\seealso{
[compute_irs_effect_sizes_simple()]
}
