% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-coverage-none.R
\name{LSMCoverage.none}
\alias{LSMCoverage.none}
\title{Set no LSMCoverage}
\usage{
\method{LSMCoverage}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for LSMCoverage
}
