% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-coverage-interface.R
\name{LSMCoverage}
\alias{LSMCoverage}
\title{Set the LSMCoverage}
\usage{
LSMCoverage(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the value of exogenous variables related to
LSMCoverage
}
