% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-effectsizes-interface.R
\name{LSMEffectSizes}
\alias{LSMEffectSizes}
\title{Set up LSM effect sizes}
\usage{
LSMEffectSizes(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up effect sizes for LSM
}
