% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-effectsizes-interface.R
\name{LSMEffectSizes}
\alias{LSMEffectSizes}
\title{Set up LSM effect sizes}
\usage{
LSMEffectSizes(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}

\item{s}{vector species index}
}
\value{
an **`xds`** object
}
\description{
Set up effect sizes for LSM
}
