% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-multiround.R
\name{make_bednet_multiround}
\alias{make_bednet_multiround}
\title{Set up dynamic forcing}
\usage{
make_bednet_multiround(
  options = list(),
  t_init = 1,
  coverage = 0.8,
  zap = 1,
  type = "pbo"
)
}
\arguments{
\item{options}{a list of options to override defaults}

\item{t_init}{the time when BedNet started}

\item{coverage}{the coverage achieved}

\item{zap}{the contact parameter}

\item{type}{the BedNet type}
}
\value{
a \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
