% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-multiround.R
\name{make_bednet_multiround}
\alias{make_bednet_multiround}
\title{Make the Multiround Bednet Coverage Object}
\usage{
make_bednet_multiround(xds_obj, options = list())
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{options}{set up options to override defaults}
}
\value{
a \strong{\code{xds}} object
}
\description{
Using information about the events (see \link{setup_bednet_events}),
including a parameter describing access, this makes a
function that computes coverage over time.
}
