% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-multiround.R
\name{make_bednet_multiround}
\alias{make_bednet_multiround}
\title{Set up dynamic forcing}
\usage{
make_bednet_multiround(xds_obj, peak, pw = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{peak}{the scaling parameter}

\item{pw}{a shape parameter}
}
\value{
an bednet multi_objage / contact model object
}
\description{
Set up a function that computes
bednet multi-round
}
