% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-multiround.R
\name{make_bednet_round}
\alias{make_bednet_round}
\title{Set up dynamic forcing}
\usage{
make_bednet_round(type, t_init, peak, contact = 1)
}
\arguments{
\item{type}{the name of the BedNet type}

\item{t_init}{the time when BedNet started}

\item{peak}{the peack access achieved}

\item{contact}{contact scaling}
}
\value{
a \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
