% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-profiles.R
\name{make_bednet_round}
\alias{make_bednet_round}
\title{Make an Bed Net Round}
\usage{
make_bednet_round(coverage_profile, start_day, peak, length = 20, pw = 1)
}
\arguments{
\item{coverage_profile}{the name of the bednet type}

\item{start_day}{the start day for the bednet round}

\item{peak}{the maximum value}

\item{length}{the number of days it took to complete spraying}

\item{pw}{a shape parameter}
}
\value{
a \code{sharkfin} function object
}
\description{
Return the parameters
to make sharkfin function for an bednet round with one of the
pesticides in the \code{coverage_profiles} table.

The parameters
specify a model for the
"effective coverage" after a mass bed net distribution
}
