% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-multiround.R
\name{make_irs_multiround}
\alias{make_irs_multiround}
\title{Set up dynamic forcing}
\usage{
make_irs_multiround(contact, xds_obj)
}
\arguments{
\item{contact}{the contact parameters to use}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
an IRS multi_objage / contact model object
}
\description{
Set up a function that computes
irs multi_objage over time.
}
