% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage-multiround.R
\name{make_irs_multiround}
\alias{make_irs_multiround}
\title{Set up dynamic forcing}
\usage{
make_irs_multiround(
  options = list(),
  t_init = 0,
  coverage = 0,
  type = "none",
  zap = 1
)
}
\arguments{
\item{options}{a list of options to override defaults}

\item{t_init}{the time when IRS started}

\item{coverage}{the coverage achieved}

\item{type}{the IRS type}

\item{zap}{the contact parameter}
}
\value{
a IRS coverage model object
}
\description{
Set up a function that computes
irs coverage over time.
}
