% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-functions.R
\name{OwnBedNets.func}
\alias{OwnBedNets.func}
\title{Set no own_bednets}
\usage{
\method{OwnBedNets}{func}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
\link{list}
}
\description{
The null model for own_bednets
}
