% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-none.R
\name{OwnBedNets.none}
\alias{OwnBedNets.none}
\title{Set no own_bednets}
\usage{
\method{OwnBedNets}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
The null model for own_bednets
}
