% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-functions.R
\name{setup_bednet_effect_sizes_func}
\alias{setup_bednet_effect_sizes_func}
\title{Set no bednet}
\usage{
setup_bednet_effect_sizes_func(xds_obj, opts)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{opts}{an optional list of arguments}
}
\value{
a bednet effect size model object
}
\description{
The null model for bednet
}
