% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-effects.R
\name{setup_bednet_effects}
\alias{setup_bednet_effects}
\title{Set Up the Bed Net Effects Object}
\usage{
setup_bednet_effects(name, xds_obj, options = list())
}
\arguments{
\item{name}{the model name}

\item{xds_obj}{an \strong{\code{xds}}  model object}

\item{options}{a list of options to override defaults}
}
\value{
the \strong{\code{xds}} model object
}
\description{
Set up an object to compute
an effect of bed nets, such as to modify
the availability of human hosts for blood
feeding
}
