% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-functions.R
\name{setup_bednet_effectsizes_func}
\alias{setup_bednet_effectsizes_func}
\title{Set no bednet}
\usage{
setup_bednet_effectsizes_func(pars, opts)
}
\arguments{
\item{pars}{an **`xds`** object}

\item{opts}{a list of options to override defaults}
}
\value{
an **`xds`** object
}
\description{
The null model for bednet
}
