% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{setup_bednet_events}
\alias{setup_bednet_events}
\title{Set Up Bed Net Evaluation}
\usage{
setup_bednet_events(
  xds_obj,
  start_day,
  net_type,
  peak_access,
  d_50 = 365,
  d_shape = 1/100,
  event_length = 20
)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{start_day}{the julian dates of bed net mass distribution events}

\item{net_type}{the type of net used}

\item{peak_access}{the fraction of the population with access to a bed net}

\item{d_50}{the half-saturation parameter}

\item{d_shape}{the access parameter}

\item{event_length}{the length of the bed net distribution event}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Set Up Bed Net Evaluation
}
