% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-interface.R
\name{setup_irs}
\alias{setup_irs}
\title{Set up dynamic forcing}
\usage{
setup_irs(
  pars,
  spray_houses_name = "none",
  spray_houses_opts = list(),
  effects_name = "none",
  effects_opts = list(),
  coverage_name = "none",
  coverage_opts = list(),
  effect_sizes_name = "none",
  effect_sizes_opts = list()
)
}
\arguments{
\item{pars}{an **`xds`** object}

\item{spray_houses_name}{the name of a model for mass bed net distribution}

\item{spray_houses_opts}{options for the bed net distribution model}

\item{effects_name}{the name of a model for bed net effects}

\item{effects_opts}{options for the bed net effects model}

\item{coverage_name}{the name of a model for bed net coverage}

\item{coverage_opts}{options for the bed net coverage model}

\item{effect_sizes_name}{the name of a model for bed net effect sizes}

\item{effect_sizes_opts}{options for the bed net effect sizes model}
}
\value{
an **`xds`** object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
