% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage-functions.R
\name{setup_irs_coverage.func}
\alias{setup_irs_coverage.func}
\title{Set up a function for IRS coverage}
\usage{
\method{setup_irs_coverage}{func}(name, pars, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an **`xds`** object}

\item{opts}{a list of options to override defaults}
}
\value{
An IRS coverage model object
}
\description{
A set up utility function for
[IRSCoverage]. Set up a s
}
