% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage-multiround.R
\name{setup_irs_coverage.multiround}
\alias{setup_irs_coverage.multiround}
\title{Set up dynamic forcing}
\usage{
\method{setup_irs_coverage}{multiround}(irs_type, pars, opts = list())
}
\arguments{
\item{pars}{an **`xds`** object}

\item{opts}{a list of options to override defaults}
}
\value{
an **`xds`** object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
