% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage-multiround.R
\name{setup_irs_coverage.multiround}
\alias{setup_irs_coverage.multiround}
\title{Set up dynamic forcing}
\usage{
\method{setup_irs_coverage}{multiround}(name, xds_obj, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{opts}{a list of options to override defaults}
}
\value{
a IRS coverage model object
}
\description{
A set up utility function for
\link{IRSCoverage}.
}
