% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-functions.R
\name{setup_irs_coverage_func}
\alias{setup_irs_coverage_func}
\title{Set up dynamic forcing}
\usage{
setup_irs_coverage_func(
  pars,
  opts = list(),
  mx = 1,
  F_season = F_flat,
  season_par = list(),
  F_trend = F_flat,
  trend_par = list()
)
}
\arguments{
\item{pars}{an **`xds`** object}

\item{opts}{a list of options to override defaults}

\item{mx}{peak irs_coverage}

\item{F_season}{a function describing a seasonal pattern over time}

\item{season_par}{an object to configure a seasonality function using [make_function]}

\item{F_trend}{a function describing a temporal trend over time}

\item{trend_par}{an object to configure a trends function using [make_function]}
}
\value{
an **`xds`** object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
