% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-interface.R
\name{setup_irs_effect_sizes}
\alias{setup_irs_effect_sizes}
\title{Set up dynamic forcing}
\usage{
setup_irs_effect_sizes(name, xds_obj, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{opts}{a list of options to override defaults}
}
\value{
a irs effect size model object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
