% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effectsizes-simple.R
\name{setup_irs_effect_sizes.simple}
\alias{setup_irs_effect_sizes.simple}
\title{Set up dynamic forcing}
\usage{
\method{setup_irs_effect_sizes}{simple}(name, xds_obj, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{opts}{a list of options to override defaults}
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
