% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-none.R
\name{setup_irs_effects.none}
\alias{setup_irs_effects.none}
\title{Set up "no irs_effects"}
\usage{
\method{setup_irs_effects}{none}(name, xds_obj, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{opts}{a list of options to override defaults}
}
\value{
a irs effects model object
}
\description{
Set up "no irs_effects"
}
