% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-multiround.R
\name{setup_irs_multiround}
\alias{setup_irs_multiround}
\title{Set up dynamic forcing}
\usage{
setup_irs_multiround(
  t_init = 1,
  coverage = 0.8,
  irs_type = "actellic",
  zap = 1,
  opts = list()
)
}
\arguments{
\item{t_init}{the time when IRS started}

\item{coverage}{the coverage achieved}

\item{irs_type}{the IRS type}

\item{zap}{the contact parameter}

\item{opts}{a list of options to override defaults}
}
\value{
an **`xds`** object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
