% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage-multiround.R
\name{setup_irs_multiround}
\alias{setup_irs_multiround}
\title{Set up dynamic forcing}
\usage{
setup_irs_multiround(
  opts = list(),
  t_init = 0,
  coverage = 0,
  type = "none",
  zap = 1
)
}
\arguments{
\item{opts}{a list of options to override defaults}

\item{t_init}{the time when IRS started}

\item{coverage}{the coverage achieved}

\item{type}{the IRS type}

\item{zap}{the contact parameter}
}
\value{
an IRS coverage model object
}
\description{
Set up a function that computes
irs coverage over time.
}
