% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-none.R
\name{setup_no_area_spray}
\alias{setup_no_area_spray}
\title{Set up no Area Spraying}
\usage{
setup_no_area_spray(pars)
}
\arguments{
\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
Set up no Area Spraying
}
