% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-interface.R
\name{setup_spray_houses}
\alias{setup_spray_houses}
\title{Set up dynamic irs}
\usage{
setup_spray_houses(name, xds_obj, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{opts}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
If dynamic irs has not
already been set up, then turn on dynamic
irs and set all the
}
