% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-none.R
\name{SprayArea.none}
\alias{SprayArea.none}
\title{Set no spray_area}
\usage{
\method{SprayArea}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
[list]
}
\description{
The null model for spray_area
}
