% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-spray_houses.R
\name{SprayHouses}
\alias{SprayHouses}
\title{Spray Houses}
\usage{
SprayHouses(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The model
}
