% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-coverage-none.R
\name{SugarBaitCoverage.none}
\alias{SugarBaitCoverage.none}
\title{No sugar baits}
\usage{
\method{SugarBaitCoverage}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
[list]
}
\description{
The null model for sugar baits
}
