% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-coverage-interface.R
\name{SugarBaitCoverage}
\alias{SugarBaitCoverage}
\title{Set the SugarBaitCoverage}
\usage{
SugarBaitCoverage(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
Sugar Baits
}
