% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar-bait-effectsizes-none.R
\name{SugarBaitEffectSizes.none}
\alias{SugarBaitEffectSizes.none}
\title{No sugar baits}
\usage{
\method{SugarBaitEffectSizes}{none}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{vector species index}
}
\value{
\link{list}
}
\description{
The null model for sugar baits
}
