% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-effectsizes-interface.R
\name{SugarBaitEffectSizes}
\alias{SugarBaitEffectSizes}
\title{Set the SugarBaitEffectSizes}
\usage{
SugarBaitEffectSizes(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the value of exogenous variables related to
Sugar Baits
}
