% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait.R
\name{SugarBaits.none}
\alias{SugarBaits.none}
\title{No sugar baits}
\usage{
\method{SugarBaits}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
\link{list}
}
\description{
The null model for sugar baits
}
