% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait.R
\name{SugarBaits}
\alias{SugarBaits}
\title{Set the SugarBaits}
\usage{
SugarBaits(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
Set the value of exogenous variables related to
Sugar Baits
}
